Convert file
