--- src/World/graveStone.d.ts
@@ -1,5 +1,0 @@
-interface Setup {
-  graveStone: {
-    create(town: Town): unknown
-  }
-}
